move command class
