+++ app/Console/Commands/MakeUserCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Console;
+namespace App\Console\Commands;
 
 use App\User;
 use Illuminate\Console\Command;
